make LoggingModel an inner class of its caller/owner
motivation:
* it being 'top level' model would potentially confuse users
* private inner classes would not be included in javadocs
